メソッド以外の関数をutilに移動
